refactor: move views out of execution
Views are used for execution state and more generally than just the
execution phase.
