+++ packages/core/src/internal/views/execution-state/get-pending-nonce-and-sender.ts
@@ -1,4 +1,7 @@
-import { ExecutionSateType, ExecutionState } from "../../types/execution-state";
+import {
+  ExecutionSateType,
+  ExecutionState,
+} from "../../execution/types/execution-state";
 
 import { getPendingOnchainInteraction } from "./get-pending-onchain-interaction";
