+++ packages/core/src/internal/views/find-result-for-future-by-id.ts
@@ -1,8 +1,8 @@
-import { SolidityParameterType } from "../../../types/module";
-import { assertIgnitionInvariant } from "../../utils/assertions";
-import { DeploymentState } from "../types/deployment-state";
-import { ExecutionResultType } from "../types/execution-result";
-import { ExecutionSateType } from "../types/execution-state";
+import { SolidityParameterType } from "../../types/module";
+import { DeploymentState } from "../execution/types/deployment-state";
+import { ExecutionResultType } from "../execution/types/execution-result";
+import { ExecutionSateType } from "../execution/types/execution-state";
+import { assertIgnitionInvariant } from "../utils/assertions";
 
 export function findResultForFutureById(
   deploymentState: DeploymentState,
