+++ packages/core/test/execution/reducers/running-a-named-library-deploy.ts
@@ -17,7 +17,7 @@ import {
   TransactionSendMessage,
 } from "../../../src/internal/execution/types/messages";
 import { NetworkInteractionType } from "../../../src/internal/execution/types/network-interaction";
-import { findExecutionStateById } from "../../../src/internal/execution/views/find-execution-state-by-id";
+import { findExecutionStateById } from "../../../src/internal/views/find-execution-state-by-id";
 import { FutureType } from "../../../src/types/module";
 
 import { applyMessages } from "./utils";
